refactor(ui): move model install error event handling to own file
No logic change.
